refactor(ruby): split call resolution and type inference into calls.go/typeinfer.go
Move call emission, target resolution, and the send-dispatch heuristic to
calls.go; the type-map engine and block-param inference to typeinfer.go.
Decompose resolveChainReceiver (chainReceiverType helper) and dispatchCall
(dispatchClassBodyCall/dispatchTopLevelCall). Collapse the duplicated
.new-assignment loop in buildLocalTypeMap/buildInstanceVarTypeMap into a
shared collectNewAssignments. Behavior-preserving: 24 goldens byte-identical.
