+++ internal/extract/ruby/ruby.go
@@ -102,15 +102,6 @@ func (Extractor) Extract(tree *sitter.Tree, source []byte, filePath string, emit
 
 func init() { extract.Register(Extractor{}) }
 
-// collectionMethods is the set of method names for which block parameter
-// type inference is supported. When a call with a block uses one of these
-// methods, the block parameter inherits the receiver's element type.
-var collectionMethods = map[string]bool{
-	"each": true, "map": true, "select": true,
-	"reject": true, "find": true, "detect": true,
-	"flat_map": true, "collect": true, "filter": true,
-}
-
 // ---- walker ----
 
 type walker struct {
@@ -179,8 +170,6 @@ func (w *walker) walk(n *sitter.Node, scope []string) error {
 // dispatchCall extracts the method name from a call node once and routes
 // to the appropriate handler. Returns true if the node was fully consumed
 // (caller should not descend into children).
-//
-//nolint:gocyclo // 27-10: retired by the ruby extractor split
 func (w *walker) dispatchCall(n *sitter.Node, scope []string) (bool, error) {
 	methodNode := n.ChildByFieldName("method")
 	if methodNode == nil {
@@ -191,42 +180,15 @@ func (w *walker) dispatchCall(n *sitter.Node, scope []string) (bool, error) {
 		return false, nil
 	}
 
-	inScope := len(scope) > 0
-
-	// Class-body-level handlers (require enclosing class/module)
-	if inScope {
-		switch methodName {
-		case "include", "extend", "prepend":
-			return false, w.emitIncludeEdge(n, scope)
-		case "has_many", "has_and_belongs_to_many", "belongs_to", "has_one":
-			return false, w.emitAssociationEdge(n, scope, methodName)
-		case "broadcasts_to", "broadcasts":
-			return false, w.emitBroadcastEdge(n, scope)
-		case "scope":
-			return false, w.emitScopeEdge(n, scope)
+	// Class-body-level handlers (require enclosing class/module) and
+	// top-level handlers (no enclosing class/module) are mutually exclusive
+	// by scope. An unrecognized method falls through to the RSpec DSL check.
+	if len(scope) > 0 {
+		if handled, err := w.dispatchClassBodyCall(n, scope, methodName); handled {
+			return false, err
 		}
-		if model.RailsCallbackNames[methodName] {
-			return false, w.emitCallbackEdges(n, scope, methodName)
-		}
-	}
-
-	// Top-level handlers (no enclosing class/module)
-	if !inScope {
-		switch methodName {
-		case "resources":
-			return true, w.handleResources(n, false)
-		case "resource":
-			return true, w.handleResources(n, true)
-		case "namespace":
-			return true, w.handleRouteNamespace(n)
-		case "pin":
-			return false, w.emitImportmapPin(n)
-		case "pin_all_from":
-			return false, w.emitImportmapPinAll(n)
-		}
-		if routeVerbs[methodName] {
-			return true, w.handleVerbRoute(n)
-		}
+	} else if consumed, handled, err := w.dispatchTopLevelCall(n, methodName); handled {
+		return consumed, err
 	}
 
 	// RSpec DSL blocks with a body — handled as test scopes.
@@ -235,6 +197,50 @@ func (w *walker) dispatchCall(n *sitter.Node, scope []string) (bool, error) {
 	}
 
 	return false, nil
+}
+
+// dispatchClassBodyCall routes a call inside a class/module body to its
+// edge emitter (include/association/broadcast/scope/callback). handled
+// reports whether the method name was recognized; class-body handlers never
+// consume the node, so dispatchCall always descends into children.
+func (w *walker) dispatchClassBodyCall(n *sitter.Node, scope []string, methodName string) (handled bool, err error) {
+	switch methodName {
+	case "include", "extend", "prepend":
+		return true, w.emitIncludeEdge(n, scope)
+	case "has_many", "has_and_belongs_to_many", "belongs_to", "has_one":
+		return true, w.emitAssociationEdge(n, scope, methodName)
+	case "broadcasts_to", "broadcasts":
+		return true, w.emitBroadcastEdge(n, scope)
+	case "scope":
+		return true, w.emitScopeEdge(n, scope)
+	}
+	if model.RailsCallbackNames[methodName] {
+		return true, w.emitCallbackEdges(n, scope, methodName)
+	}
+	return false, nil
+}
+
+// dispatchTopLevelCall routes a call at file scope (route DSL, importmap pin)
+// to its handler. handled reports whether the method name was recognized;
+// consumed reports whether the node was fully handled (route blocks consume,
+// importmap pins do not).
+func (w *walker) dispatchTopLevelCall(n *sitter.Node, methodName string) (consumed, handled bool, err error) {
+	switch methodName {
+	case "resources":
+		return true, true, w.handleResources(n, false)
+	case "resource":
+		return true, true, w.handleResources(n, true)
+	case "namespace":
+		return true, true, w.handleRouteNamespace(n)
+	case "pin":
+		return false, true, w.emitImportmapPin(n)
+	case "pin_all_from":
+		return false, true, w.emitImportmapPinAll(n)
+	}
+	if routeVerbs[methodName] {
+		return true, true, w.handleVerbRoute(n)
+	}
+	return false, false, nil
 }
 
 func (w *walker) walkChildren(n *sitter.Node, scope []string) error {
@@ -290,680 +296,6 @@ func isInsideNestedBlock(n, parent *sitter.Node) bool {
 		}
 	}
 	return false
-}
-
-// extractBlockParams pulls simple identifier parameter names from a
-// block_parameters node.
-//
-// Returns nil when:
-//   - the block has no parameters node (e.g. bare block without |...|)
-//   - any parameter is not a simple identifier (destructuring, splat,
-//     optional, etc.)
-//
-// In both cases the caller should skip block parameter type inference
-// and fall back to walking the block body with the original local type
-// map.
-func extractBlockParams(block *sitter.Node, source []byte) []string {
-	paramsNode := block.ChildByFieldName("parameters")
-	if paramsNode == nil {
-		return nil
-	}
-	var params []string
-	count := paramsNode.NamedChildCount()
-	for i := uint(0); i < count; i++ {
-		c := paramsNode.NamedChild(i)
-		if c == nil {
-			continue
-		}
-		switch c.Kind() {
-		case "identifier":
-			params = append(params, extract.Text(c, source))
-		default:
-			// Destructuring, splat, optional parameters — skip the whole block.
-			return nil
-		}
-	}
-	return params
-}
-
-// extractElementType extracts the element type from a collection type string.
-// Handles Array[Type] syntax and falls back to a plural→singular heuristic.
-func extractElementType(collectionType string) string {
-	if collectionType == "" {
-		return ""
-	}
-	// Array[Order] → Order
-	if strings.HasPrefix(collectionType, "Array[") && strings.HasSuffix(collectionType, "]") {
-		return collectionType[6 : len(collectionType)-1]
-	}
-	// Plural → singular heuristic (e.g. orders → Order, users → User)
-	singular := singularize(collectionType)
-	if singular != collectionType {
-		return pascalCase(singular)
-	}
-	return collectionType
-}
-
-// mergeMaps returns a new map containing all entries from base, overlaid
-// with entries from overlay. Neither input map is modified.
-func mergeMaps(base, overlay map[string]string) map[string]string {
-	result := make(map[string]string, len(base)+len(overlay))
-	for k, v := range base {
-		result[k] = v
-	}
-	for k, v := range overlay {
-		result[k] = v
-	}
-	return result
-}
-
-// emitCall produces a calls edge for one `call` node. The target is
-// resolved from the receiver when possible: `self` and implicit calls
-// are emitted as `self.name` so the resolver rewrites them to the
-// enclosing class; constant receivers are emitted as `Const.name` for
-// exact matching; local-variable receivers are resolved via a lightweight
-// intra-method type map built from `X = Class.new` assignments; method
-// chains are stripped to the trailing method name. `send` /
-// `public_send` / `__send__` with a literal symbol or string first
-// argument is emitted with confidence 0.7; anything else in that family
-// is skipped.
-func (w *walker) emitCall(n *sitter.Node, source string, scope []string, localTypes, ivarTypes map[string]string) error {
-	methodNode := n.ChildByFieldName("method")
-	if methodNode == nil {
-		return nil
-	}
-	methodName := extract.Text(methodNode, w.source)
-	if methodName == "" {
-		return nil
-	}
-
-	switch methodName {
-	case "send", "public_send", "__send__":
-		target, ok := literalSendTarget(n, w.source)
-		if ok {
-			line := extract.Line(n.StartPosition())
-			return w.emit.Edge(extract.EmittedEdge{
-				SourceQualified: source,
-				TargetQualified: target,
-				Kind:            model.EdgeCalls,
-				Line:            &line,
-				Confidence:      extract.ConfidenceDynamic,
-			})
-		}
-		// Heuristic: variable-based dynamic dispatch on self.
-		recv := n.ChildByFieldName("receiver")
-		if recv == nil || recv.Kind() == "self" {
-			if target, conf, ok := inferSendTargetFromVariable(n, w.source); ok {
-				line := extract.Line(n.StartPosition())
-				return w.emit.Edge(extract.EmittedEdge{
-					SourceQualified: source,
-					TargetQualified: target,
-					Kind:            model.EdgeCalls,
-					Line:            &line,
-					Confidence:      conf,
-				})
-			}
-		}
-		return nil
-	}
-
-	recv := n.ChildByFieldName("receiver")
-	target, confidence := w.resolveCallTarget(recv, methodName, scope, localTypes, ivarTypes)
-	if target == "" {
-		return nil
-	}
-	return w.emitCallWithConfidence(n, source, scope, localTypes, ivarTypes, confidence)
-}
-
-// emitCallWithConfidence is emitCall's core logic with an injectable
-// confidence value. Used by both production-method emission (1.0 / 0.7)
-// and test-block emission (0.8).
-func (w *walker) emitCallWithConfidence(n *sitter.Node, source string, scope []string, localTypes, ivarTypes map[string]string, confidence float64) error {
-	methodNode := n.ChildByFieldName("method")
-	if methodNode == nil {
-		return nil
-	}
-	methodName := extract.Text(methodNode, w.source)
-	if methodName == "" {
-		return nil
-	}
-	line := extract.Line(n.StartPosition())
-
-	switch methodName {
-	case "send", "public_send", "__send__":
-		target, ok := literalSendTarget(n, w.source)
-		if ok {
-			return w.emit.Edge(extract.EmittedEdge{
-				SourceQualified: source,
-				TargetQualified: target,
-				Kind:            model.EdgeCalls,
-				Line:            &line,
-				Confidence:      confidence,
-			})
-		}
-		// Heuristic: variable-based dynamic dispatch on self.
-		recv := n.ChildByFieldName("receiver")
-		if recv == nil || recv.Kind() == "self" {
-			if target, conf, ok := inferSendTargetFromVariable(n, w.source); ok {
-				return w.emit.Edge(extract.EmittedEdge{
-					SourceQualified: source,
-					TargetQualified: target,
-					Kind:            model.EdgeCalls,
-					Line:            &line,
-					Confidence:      conf,
-				})
-			}
-		}
-		return nil
-	}
-
-	recv := n.ChildByFieldName("receiver")
-	target, _ := w.resolveCallTarget(recv, methodName, scope, localTypes, ivarTypes)
-	if target == "" {
-		return nil
-	}
-	if err := w.emit.Edge(extract.EmittedEdge{
-		SourceQualified: source,
-		TargetQualified: target,
-		Kind:            model.EdgeCalls,
-		Line:            &line,
-		Confidence:      confidence,
-	}); err != nil {
-		return err
-	}
-
-	// If this call has a block, infer block parameter types from the
-	// receiver's collection type and walk the block body with an
-	// augmented local type map. When inference is not possible (non-
-	// collection method, destructuring params, unknown receiver type),
-	// we still walk the block so calls inside are not lost.
-	if block := n.ChildByFieldName("block"); block != nil {
-		paramTypes := w.inferBlockParamTypes(n, scope, localTypes, ivarTypes)
-		blockTypes := localTypes
-		if paramTypes != nil {
-			blockTypes = mergeMaps(localTypes, paramTypes)
-		}
-		if err := extract.WalkNamedDescendants(block, "call", func(c *sitter.Node) error {
-			if isInsideNestedBlock(c, block) {
-				return nil
-			}
-			return w.emitCall(c, source, scope, blockTypes, ivarTypes)
-		}); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// coreNoiseMethods are ubiquitous core-Ruby / ActiveSupport methods
-// whose bare name (receiver type unknown) matches far too many unrelated
-// symbols. When the receiver type cannot be inferred we drop the call
-// edge rather than let the resolver's unqualified fallback bind it to an
-// arbitrary same-named method elsewhere — e.g. `count.zero?` resolving
-// to a `Money.zero` singleton. A missing low-confidence edge is cheaper
-// than a false caller that inflates blast radius. The ERB analogue, for
-// framework context accessors referenced bare in a template, is
-// erbHelperSkip in internal/extract/erb/erb.go.
-var coreNoiseMethods = map[string]bool{
-	"zero?": true, "positive?": true, "negative?": true, "nonzero?": true,
-	"present?": true, "blank?": true, "nil?": true, "empty?": true,
-	"to_s": true, "to_i": true, "to_a": true, "to_h": true, "to_sym": true,
-	"freeze": true, "dup": true, "clone": true, "presence": true, "call": true,
-	// Object/reflection methods: a bare `x.is_a?` / `x.respond_to?` on an
-	// unknown receiver otherwise binds to a coincidental same-named app symbol
-	// (e.g. a test fake's `#is_a?`). Never a meaningful caller edge.
-	"is_a?": true, "kind_of?": true, "instance_of?": true, "respond_to?": true,
-	"frozen?": true, "itself": true, "inspect": true, "object_id": true,
-}
-
-// unresolvedCall is the emit decision for a call whose receiver type is
-// unknown: drop ubiquitous core-method names, otherwise emit the bare
-// name at ConfidenceUnresolved for the resolver's unqualified fallback.
-func unresolvedCall(methodName string) (string, float64) {
-	if coreNoiseMethods[methodName] {
-		return "", 0
-	}
-	return methodName, extract.ConfidenceUnresolved
-}
-
-// resolveCallTarget decides what target string to emit for a call node.
-// It returns the target string and the confidence level.
-func (w *walker) resolveCallTarget(recv *sitter.Node, methodName string, scope []string, localTypes, ivarTypes map[string]string) (string, float64) {
-	if recv == nil {
-		return "self." + methodName, 1.0
-	}
-
-	switch recv.Kind() {
-	case "self":
-		return "self." + methodName, 1.0
-	case "constant", "scope_resolution":
-		if recvText := strings.TrimSpace(extract.Text(recv, w.source)); recvText != "" {
-			return recvText + "." + methodName, 1.0
-		}
-	case "identifier":
-		name := extract.Text(recv, w.source)
-		if typ, ok := localTypes[name]; ok {
-			return typ + "#" + methodName, extract.ConfidenceDynamic
-		}
-		return unresolvedCall(methodName)
-	case "instance_variable":
-		name := extract.Text(recv, w.source)
-		if typ, ok := localTypes[name]; ok {
-			return typ + "#" + methodName, extract.ConfidenceDynamic
-		}
-		if typ, ok := ivarTypes[name]; ok {
-			return typ + "#" + methodName, extract.ConfidenceDynamic
-		}
-		return unresolvedCall(methodName)
-	case "call":
-		// Method chain — strip to the trailing method unless the inner
-		// call is `.new` on a constant or `self`, in which case we can
-		// infer the result type is an instance of that class.
-		if typ := typeFromNewCall(recv, w.source, scope); typ != "" {
-			return typ + "#" + methodName, extract.ConfidenceDynamic
-		}
-		// Try to resolve multi-hop chains via return-type map.
-		if typ := w.resolveChainReceiver(recv, scope, localTypes, 1); typ != "" {
-			return typ + "#" + methodName, extract.ConfidenceDynamic
-		}
-		return unresolvedCall(methodName)
-	}
-
-	return unresolvedCall(methodName)
-}
-
-// resolveChainReceiver recursively resolves a call-chain receiver to a type
-// by looking up local-variable types and method return types. It caps at 3
-// hops to avoid exponential lookups and absurd qualified names.
-//
-//nolint:gocyclo // 27-10: retired by the ruby extractor split
-func (w *walker) resolveChainReceiver(recv *sitter.Node, scope []string, localTypes map[string]string, depth int) string {
-	if depth > 3 || recv == nil || recv.Kind() != "call" {
-		return ""
-	}
-	methodNode := recv.ChildByFieldName("method")
-	if methodNode == nil {
-		return ""
-	}
-	methodName := extract.Text(methodNode, w.source)
-	innerRecv := recv.ChildByFieldName("receiver")
-
-	// Case 1: receiver is a local variable with known type.
-	if innerRecv != nil && innerRecv.Kind() == "identifier" {
-		if typ, ok := localTypes[extract.Text(innerRecv, w.source)]; ok {
-			qualified := typ + "#" + methodName
-			if ret, ok := w.returnTypes[qualified]; ok {
-				return ret
-			}
-		}
-	}
-
-	// Case 2: receiver is self (implicit or explicit).
-	if innerRecv == nil || innerRecv.Kind() == "self" {
-		parent := strings.Join(scope, "::")
-		qualified := parent + "#" + methodName
-		if ret, ok := w.returnTypes[qualified]; ok {
-			return ret
-		}
-	}
-
-	// Case 3: receiver is another chain (recursive).
-	if innerRecv != nil && innerRecv.Kind() == "call" {
-		innerType := w.resolveChainReceiver(innerRecv, scope, localTypes, depth+1)
-		if innerType != "" {
-			qualified := innerType + "#" + methodName
-			if ret, ok := w.returnTypes[qualified]; ok {
-				return ret
-			}
-		}
-	}
-
-	return ""
-}
-
-// inferBlockParamTypes determines the element type for block parameters
-// when a call with a block uses a known collection method (each, map, etc.).
-// It looks up the receiver's collection type from local variables, instance
-// variables, or method chains, then extracts the singular element type.
-// Returns nil when the method is not in the whitelist or the type cannot
-// be inferred.
-func (w *walker) inferBlockParamTypes(callNode *sitter.Node, scope []string, localTypes, ivarTypes map[string]string) map[string]string {
-	methodNode := callNode.ChildByFieldName("method")
-	if methodNode == nil {
-		return nil
-	}
-	methodName := extract.Text(methodNode, w.source)
-	if !collectionMethods[methodName] {
-		return nil
-	}
-
-	recv := callNode.ChildByFieldName("receiver")
-	if recv == nil {
-		return nil
-	}
-
-	var collectionType string
-	switch recv.Kind() {
-	case "identifier":
-		collectionType = localTypes[extract.Text(recv, w.source)]
-	case "instance_variable":
-		name := extract.Text(recv, w.source)
-		if typ, ok := localTypes[name]; ok {
-			collectionType = typ
-		} else if typ, ok := ivarTypes[name]; ok {
-			collectionType = typ
-		}
-	case "call":
-		// For chain resolution, e.g. user.orders.each { |order| ... }
-		collectionType = w.resolveChainReceiver(recv, scope, localTypes, 1)
-	}
-
-	if collectionType == "" {
-		return nil
-	}
-
-	elementType := extractElementType(collectionType)
-	if elementType == "" {
-		return nil
-	}
-
-	block := callNode.ChildByFieldName("block")
-	if block == nil {
-		return nil
-	}
-	params := extractBlockParams(block, w.source)
-	if params == nil {
-		return nil
-	}
-
-	result := make(map[string]string)
-	for _, param := range params {
-		result[param] = elementType
-	}
-	return result
-}
-
-// buildLocalTypeMap scans a method body for local-variable assignments
-// whose RHS is `ClassName.new(...)` and returns a map from variable name
-// to class name. This enables lightweight receiver resolution for the
-// most common object-creation pattern in Ruby.
-func buildLocalTypeMap(body *sitter.Node, source []byte) map[string]string {
-	types := make(map[string]string)
-	if body == nil {
-		return types
-	}
-	for _, kind := range []string{"assignment", "operator_assignment"} {
-		_ = extract.WalkNamedDescendants(body, kind, func(n *sitter.Node) error {
-			lhs := n.ChildByFieldName("left")
-			rhs := n.ChildByFieldName("right")
-			if lhs == nil || rhs == nil || lhs.Kind() != "identifier" {
-				return nil
-			}
-			if typ := typeFromNewCall(rhs, source, nil); typ != "" {
-				types[extract.Text(lhs, source)] = typ
-			}
-			return nil
-		})
-	}
-	return types
-}
-
-// addRescueBindings records the type of each `rescue ExceptionType => var`
-// clause in the method body into localTypes, so a call on the bound variable
-// (`raise if e.retriable?`) resolves to ExceptionType#method instead of the
-// resolver's bare-name fallback, which would otherwise bind it to an arbitrary
-// same-named method on an unrelated class.
-//
-// Only single-type rescues with an identifier variable are recorded:
-// `rescue A, B => e` leaves the variable's type ambiguous, and a bare
-// `rescue => e` has no type. An existing entry (e.g. an explicit
-// `X = Class.new`) is never overwritten. The map is method-wide rather than
-// rescue-scoped, so two rescues binding the same name to different types are
-// last-write-wins — rare, and a missed/loose bind only costs one weak edge.
-func (w *walker) addRescueBindings(body *sitter.Node, localTypes map[string]string) {
-	if body == nil {
-		return
-	}
-	_ = extract.WalkNamedDescendants(body, "rescue", func(r *sitter.Node) error {
-		typ := w.singleRescueType(r)
-		if typ == "" {
-			return nil
-		}
-		varName := rescueVariableName(r, w.source)
-		if varName == "" {
-			return nil
-		}
-		if _, exists := localTypes[varName]; !exists {
-			localTypes[varName] = typ
-		}
-		return nil
-	})
-}
-
-// singleRescueType returns the fully-qualified exception type of a rescue
-// clause when it names exactly one constant/scope_resolution, resolving the
-// trailing segment through pkgBindings so `rescue ApiError` inside its own
-// namespace still yields the qualified name. Returns "" for multi-type or
-// typeless rescues.
-func (w *walker) singleRescueType(rescueNode *sitter.Node) string {
-	var exceptions *sitter.Node
-	for i := uint(0); i < rescueNode.NamedChildCount(); i++ {
-		if c := rescueNode.NamedChild(i); c.Kind() == "exceptions" {
-			exceptions = c
-			break
-		}
-	}
-	if exceptions == nil || exceptions.NamedChildCount() != 1 {
-		return ""
-	}
-	typeNode := exceptions.NamedChild(0)
-	switch typeNode.Kind() {
-	case "constant", "scope_resolution":
-	default:
-		return ""
-	}
-	text := strings.TrimSpace(extract.Text(typeNode, w.source))
-	if text == "" {
-		return ""
-	}
-	if q, ok := w.pkgBindings[lastConstSegment(text)]; ok {
-		return q
-	}
-	return text
-}
-
-// rescueVariableName returns the identifier bound by a rescue clause's
-// `=> var`, or "" when the binding is absent or not a simple identifier.
-func rescueVariableName(rescueNode *sitter.Node, source []byte) string {
-	for i := uint(0); i < rescueNode.NamedChildCount(); i++ {
-		c := rescueNode.NamedChild(i)
-		if c.Kind() != "exception_variable" {
-			continue
-		}
-		for j := uint(0); j < c.NamedChildCount(); j++ {
-			if v := c.NamedChild(j); v.Kind() == "identifier" {
-				return extract.Text(v, source)
-			}
-		}
-	}
-	return ""
-}
-
-// lastConstSegment returns the trailing `::`-separated segment of a constant
-// reference: "SocialCommerce::Meta::ApiError" → "ApiError", "ApiError" → "ApiError".
-func lastConstSegment(name string) string {
-	if i := strings.LastIndex(name, "::"); i >= 0 {
-		return name[i+len("::"):]
-	}
-	return name
-}
-
-// buildInstanceVarTypeMap scans a class body for `initialize` methods and
-// looks for instance-variable assignments whose RHS is `ClassName.new(...)`.
-// Returns a map from @ivar_name to class_name.
-func buildInstanceVarTypeMap(body *sitter.Node, source []byte) map[string]string {
-	types := make(map[string]string)
-	if body == nil {
-		return types
-	}
-	_ = extract.WalkNamedDescendants(body, "method", func(n *sitter.Node) error {
-		nameNode := n.ChildByFieldName("name")
-		if nameNode == nil || extract.Text(nameNode, source) != "initialize" {
-			return nil
-		}
-		initBody := n.ChildByFieldName("body")
-		if initBody == nil {
-			return nil
-		}
-		for _, kind := range []string{"assignment", "operator_assignment"} {
-			_ = extract.WalkNamedDescendants(initBody, kind, func(a *sitter.Node) error {
-				lhs := a.ChildByFieldName("left")
-				rhs := a.ChildByFieldName("right")
-				if lhs == nil || rhs == nil || lhs.Kind() != "instance_variable" {
-					return nil
-				}
-				if typ := typeFromNewCall(rhs, source, nil); typ != "" {
-					types[extract.Text(lhs, source)] = typ
-				}
-				return nil
-			})
-		}
-		return nil
-	})
-	return types
-}
-
-// typeFromNewCall returns the receiver class name when the node is a
-// call to `.new`, e.g. `TopicCreator.new(...)` → `"TopicCreator"`.
-// When the receiver is `self`, the enclosing class scope is used.
-func typeFromNewCall(n *sitter.Node, source []byte, scope []string) string {
-	if n == nil || n.Kind() != "call" {
-		return ""
-	}
-	methodNode := n.ChildByFieldName("method")
-	if methodNode == nil || extract.Text(methodNode, source) != "new" {
-		return ""
-	}
-	recv := n.ChildByFieldName("receiver")
-	if recv == nil {
-		return ""
-	}
-	switch recv.Kind() {
-	case "constant", "scope_resolution":
-		return extract.Text(recv, source)
-	case "self":
-		return strings.Join(scope, "::")
-	}
-	return ""
-}
-
-// buildReturnTypeMap scans a class/module body for methods whose body is
-// a single expression or a single `return` statement that calls `Class.new(...)`.
-// It returns a map from qualified method name to the inferred class name.
-func buildReturnTypeMap(body *sitter.Node, source []byte, scope []string) map[string]string {
-	types := make(map[string]string)
-	if body == nil {
-		return types
-	}
-	parent := strings.Join(scope, "::")
-	_ = extract.WalkNamedDescendants(body, "method", func(n *sitter.Node) error {
-		return recordMethodReturnType(n, source, parent, types, false)
-	})
-	_ = extract.WalkNamedDescendants(body, "singleton_method", func(n *sitter.Node) error {
-		return recordMethodReturnType(n, source, parent, types, true)
-	})
-	return types
-}
-
-// recordMethodReturnType extracts the return type from a single method
-// and records it in the map if it matches the simple factory pattern.
-func recordMethodReturnType(n *sitter.Node, source []byte, parent string, types map[string]string, singleton bool) error {
-	nameNode := n.ChildByFieldName("name")
-	if nameNode == nil {
-		return nil
-	}
-	name := extract.Text(nameNode, source)
-	if name == "" {
-		return nil
-	}
-	var qualified string
-	switch {
-	case parent == "":
-		qualified = name
-	case singleton:
-		qualified = parent + "." + name
-	default:
-		qualified = parent + "#" + name
-	}
-	body := n.ChildByFieldName("body")
-	if body == nil {
-		return nil
-	}
-	// Look for a single expression or single return statement.
-	var returnExpr *sitter.Node
-	childCount := body.NamedChildCount()
-	if childCount == 1 {
-		child := body.NamedChild(0)
-		if child.Kind() == "return" {
-			// return node has an argument_list as its first named child
-			if argList := child.NamedChild(0); argList != nil && argList.Kind() == "argument_list" {
-				if argList.NamedChildCount() == 1 {
-					returnExpr = argList.NamedChild(0)
-				}
-			}
-		} else {
-			returnExpr = child
-		}
-	}
-	if returnExpr != nil {
-		if typ := typeFromNewCall(returnExpr, source, nil); typ != "" {
-			types[qualified] = typ
-		}
-	}
-	return nil
-}
-
-// buildFileReturnTypeMap walks the entire AST and builds a map of all
-// method qualified names → return types for simple factory methods.
-func buildFileReturnTypeMap(root *sitter.Node, source []byte) map[string]string {
-	types := make(map[string]string)
-	var walk func(n *sitter.Node, scope []string)
-	walk = func(n *sitter.Node, scope []string) {
-		if n == nil {
-			return
-		}
-		switch n.Kind() {
-		case "class", "module":
-			nameNode := n.ChildByFieldName("name")
-			if nameNode != nil {
-				name := extract.Text(nameNode, source)
-				if name != "" {
-					segments := strings.Split(name, "::")
-					newScope := append(slices.Clone(scope), segments...)
-					body := n.ChildByFieldName("body")
-					if body != nil {
-						classTypes := buildReturnTypeMap(body, source, newScope)
-						for k, v := range classTypes {
-							types[k] = v
-						}
-						// Continue walking children with new scope.
-						count := body.NamedChildCount()
-						for i := uint(0); i < count; i++ {
-							walk(body.NamedChild(i), newScope)
-						}
-					}
-					return
-				}
-			}
-		}
-		// For non-class/module nodes, walk children with current scope.
-		count := n.NamedChildCount()
-		for i := uint(0); i < count; i++ {
-			walk(n.NamedChild(i), scope)
-		}
-	}
-	walk(root, nil)
-	return types
 }
 
 // statementParents is the set of tree-sitter-ruby node kinds whose
@@ -1139,149 +471,6 @@ func collectLocalNames(body *sitter.Node, source []byte, seed map[string]bool) m
 		return nil
 	})
 	return locals
-}
-
-// literalSendTarget extracts the target name from `send(:name)` /
-// `public_send("name")` / `__send__(:name)` when the first argument is
-// a bare symbol or string literal. Everything else is unresolvable.
-// The tree-sitter-ruby grammar exposes a string's payload as a named
-// `string_content` child (not a named field), and a symbol node
-// carries a leading colon; both are looked up structurally. If the
-// grammar shape drifts, we return false visibly rather than falling
-// back to quote stripping — explicit failure beats degraded output.
-func literalSendTarget(call *sitter.Node, source []byte) (string, bool) {
-	args := call.ChildByFieldName("arguments")
-	if args == nil || args.NamedChildCount() == 0 {
-		return "", false
-	}
-	first := args.NamedChild(0)
-	if first == nil {
-		return "", false
-	}
-	switch first.Kind() {
-	case "simple_symbol":
-		return strings.TrimPrefix(extract.Text(first, source), ":"), true
-	case "string":
-		count := first.NamedChildCount()
-		for i := uint(0); i < count; i++ {
-			c := first.NamedChild(i)
-			if c != nil && c.Kind() == "string_content" {
-				return extract.Text(c, source), true
-			}
-		}
-	}
-	return "", false
-}
-
-// methodNamePatterns lists variable-name substrings that suggest the
-// variable holds a method name. Used by the dynamic-dispatch heuristic
-// to avoid emitting edges for every variable-based send() call.
-var methodNamePatterns = []string{
-	"callback", "handler", "method", "action", "hook",
-	"event", "listener", "processor", "task", "job",
-	"name", "attr",
-}
-
-// ConfidenceHeuristicDispatch is the confidence for variable-inferred
-// dynamic dispatch edges. Very low — we're guessing the method name from
-// a variable assignment, which could be wrong.
-const ConfidenceHeuristicDispatch = extract.ConfidenceUnresolved / 2
-
-// findEnclosingMethodBody walks up from n to the nearest "method" node
-// and returns its "body" child, or nil if none is found.
-func findEnclosingMethodBody(n *sitter.Node) *sitter.Node {
-	for p := n.Parent(); p != nil; p = p.Parent() {
-		if p.Kind() == "method" || p.Kind() == "singleton_method" {
-			return p.ChildByFieldName("body")
-		}
-	}
-	return nil
-}
-
-// traceVariableAssignment scans body for assignments to varName that
-// appear before the given call node, and returns the literal symbol or
-// string value from the RHS of the last such assignment. It only looks
-// at direct children (not nested blocks or methods) to keep the heuristic
-// simple and fast.
-func traceVariableAssignment(body *sitter.Node, varName string, source []byte, call *sitter.Node) (string, bool) {
-	if body == nil || call == nil {
-		return "", false
-	}
-	callRow := call.StartPosition().Row
-	var result string
-	found := false
-	for _, kind := range []string{"assignment", "operator_assignment"} {
-		_ = extract.WalkNamedDescendants(body, kind, func(n *sitter.Node) error {
-			// Skip assignments that appear after the send call.
-			if n.StartPosition().Row > callRow {
-				return nil
-			}
-			lhs := n.ChildByFieldName("left")
-			rhs := n.ChildByFieldName("right")
-			if lhs == nil || rhs == nil {
-				return nil
-			}
-			if lhs.Kind() != "identifier" || extract.Text(lhs, source) != varName {
-				return nil
-			}
-			switch rhs.Kind() {
-			case "simple_symbol":
-				result = strings.TrimPrefix(extract.Text(rhs, source), ":")
-				found = true
-			case "string":
-				count := rhs.NamedChildCount()
-				for i := uint(0); i < count; i++ {
-					c := rhs.NamedChild(i)
-					if c != nil && c.Kind() == "string_content" {
-						result = extract.Text(c, source)
-						found = true
-						break
-					}
-				}
-			}
-			return nil
-		})
-	}
-	return result, found
-}
-
-// inferSendTargetFromVariable applies a heuristic to variable-based
-// dynamic dispatch: if the first argument to send/public_send/__send__
-// is an identifier whose name suggests a method name, and we can trace
-// the variable back to a literal symbol or string assignment, return
-// the inferred target with very low confidence.
-func inferSendTargetFromVariable(call *sitter.Node, source []byte) (string, float64, bool) {
-	args := call.ChildByFieldName("arguments")
-	if args == nil || args.NamedChildCount() == 0 {
-		return "", 0, false
-	}
-	first := args.NamedChild(0)
-	if first == nil || first.Kind() != "identifier" {
-		return "", 0, false
-	}
-	varName := extract.Text(first, source)
-
-	// Only apply heuristic when the variable name suggests a method name.
-	lowerVar := strings.ToLower(varName)
-	matchesPattern := false
-	for _, pat := range methodNamePatterns {
-		if strings.Contains(lowerVar, pat) {
-			matchesPattern = true
-			break
-		}
-	}
-	if !matchesPattern {
-		return "", 0, false
-	}
-
-	body := findEnclosingMethodBody(call)
-	if body == nil {
-		return "", 0, false
-	}
-	if target, ok := traceVariableAssignment(body, varName, source, call); ok {
-		return target, ConfidenceHeuristicDispatch, true
-	}
-	return "", 0, false
 }
 
 // collectConstants recursively pre-scans for constant assignments so
